[CHORE] 폴더 정리
