+++ app/src/main/java/com/example/androidseminar/data/Info.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar
+package com.example.androidseminar.data
 
 data class Info(
     val followerImgUrl: String,
