+++ app/src/main/java/com/example/androidseminar/data/RequestLoginData.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar
+package com.example.androidseminar.data
 
 data class RequestLoginData(
     val email: String,
