+++ app/src/main/java/com/example/androidseminar/data/ResponseSignUpData.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar
+package com.example.androidseminar.data
 
 data class ResponseSignUpData(
     val status: Int,
