+++ app/src/main/java/com/example/androidseminar/fragments/HomeFollowingFragment.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar
+package com.example.androidseminar.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
