+++ app/src/main/java/com/example/androidseminar/fragments/ImageFragment.kt
@@ -1,28 +1,18 @@
-package com.example.androidseminar
+package com.example.androidseminar.fragments
 
 import android.Manifest
-import android.app.Activity
-import android.app.Activity.RESULT_CANCELED
-import android.app.Activity.RESULT_OK
 import android.content.Intent
 import android.content.pm.PackageManager
-import android.graphics.ImageDecoder
-import android.os.Build
 import android.os.Bundle
-import android.provider.MediaStore
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.Toast
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import com.bumptech.glide.Glide
-import com.example.androidseminar.databinding.FragmentFollowerBinding
 import com.example.androidseminar.databinding.FragmentImageBinding
 
 class ImageFragment : Fragment() {
