+++ app/src/main/java/com/example/androidseminar/util/ItemActionListener.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar
+package com.example.androidseminar.util
 
 interface ItemActionListener {
     fun onItemMoved(from: Int, to: Int)
