+++ app/src/main/java/com/example/androidseminar/util/NestedScrollableHost.kt
@@ -1,4 +1,4 @@
-package com.example.androidseminar
+package com.example.androidseminar.util
 
 import android.content.Context
 import android.util.AttributeSet
