+++ app/src/main/res/layout/activity_home.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".adapter.HomeActivity">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/home_vp"
