+++ app/src/main/res/layout/fragment_follower.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FollowerFragment">
+    tools:context=".fragments.FollowerFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/follower_recyclerview"
