+++ app/src/main/res/layout/fragment_home_follower.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="#dfdfdf"
     android:layout_height="match_parent"
-    tools:context=".HomeFollowerFragment">
+    tools:context=".fragments.HomeFollowerFragment">
 
     <TextView
         android:layout_width="wrap_content"
