+++ app/src/main/res/layout/fragment_image.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ImageFragment">
+    tools:context=".fragments.ImageFragment">
 
     <ImageView
         android:id="@+id/image_iv"
