+++ app/src/main/res/layout/item_follower.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="followerItem"
-            type="com.example.androidseminar.Info" />
+            type="com.example.androidseminar.data.Info" />
 
     </data>
